some modifications in VS.NET project
git-svn-id: http://crashrpt.googlecode.com/svn@139 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/CodeReview/src/crashsender/src/ProgressDlg.cpp
@@ -48,7 +48,7 @@ void CProgressDlg::Start()
 
 LRESULT CProgressDlg::OnTimer(UINT /*uMsg*/, WPARAM wParam, LPARAM /*lParam*/, BOOL& /*bHandled*/)
 {
-  WORD wTimerId = wParam;
+  WORD wTimerId = (WORD)wParam;
 
   if(wTimerId==0)
   {
